Working on changing the way `all.dvi' is created.  Making it possible to
generate it so that each of the sections either starts at page 1 for printing them
out separately or so that the page numbers are continuous.  In this case, a table
of contents should also be produced.

Now, TeX is always run on `all.tex', which is _not_ a generated file.  It includes
all the generated ones.  To get separate documents, I plan to use `dviselect' to
extract them from `all.dvi'.