This version works.  I simplified the code for creating `sep_all.sh',
`all_sep.*' and `all_no_sep.*'.  Now using `tr' to add backslashes as continuation
characters at the ends of lines in `sep_all.sh'.  Now printing `|' characters instead
of backslashes to `XXX.sh', which is used to create `sep_all.sh'.  It is therefore no
longer necessary to change the catcode of `\' in `all.tex'.