In `main' (in `songlist.cxx'): Bug fix: If there was no comma
in |iter->second|, e.g., if it was "None" or "Unknown", the vertical
spacing between the corresponding line and the following line was too
small.  In this case, an empty \vbox with the depth of `\commabox'
(defined in `songlist.mac') is inserted at the beginning of the line
using `\vtop'.