Added code for writing "new" filecards to `all_no_sep.pdf' and `all_sep.pdf'.  That is,
filecards created since a date specified using the `filecard-date' command-line option.

This option takes an optional argument in the form of a "date string".  This is a string that
is a valid argument to the `date' shell command.  In addition, it can be in the form
"2021.09.03", i.e., with periods, which is not accepted by `date'.

If the option is used without an argument, the default is used, namely `-1week'.  This means
that "new filecards" will be made for songs that have been added to the database in the past week.