Added database entries for "Jet Song".
BUG FIX: In `process_tocs_and_npt': For songs with no lead sheets that
are sorted by production, they are no longer printed to the lead sheet
list files.  They are printed to the file cards.  That is, a filecard
is made for the individual songs and the songs appear on the filecard
for the production, whether there are lead sheets or not.  This is the
way it's supposed to work.